+++ assets/src/components/editing/elements/youtube/YoutubeModal.tsx
@@ -5,30 +5,33 @@ import * as ContentModel from 'data/content/model/elements/types';
 interface ModalProps {
   onDone: (x: any) => void;
   onCancel: () => void;
-  model: ContentModel.Image;
+  model: ContentModel.YouTube;
 }
-export const ImageModal = ({ onDone, onCancel, model }: ModalProps) => {
+export const YouTubeModal = ({ onDone, onCancel, model }: ModalProps) => {
+  const [src, setSrc] = useState(model.src);
   const [alt, setAlt] = useState(model.alt);
   const [width, setWidth] = useState(model.width);
 
   return (
-    <FullScreenModal onCancel={(_e) => onCancel()} onDone={() => onDone({ alt, width })}>
+    <FullScreenModal onCancel={(_e) => onCancel()} onDone={() => onDone({ alt, width, src })}>
       <div>
         <h3 className="mb-2">Settings</h3>
-        <div
-          className="mx-auto mb-4"
-          style={{
-            width: 300,
-            height: 200,
-            backgroundImage: `url(${model.src})`,
-            backgroundPosition: '50% 50%',
-            backgroundSize: 'contain',
-            backgroundRepeat: 'no-repeat',
-          }}
-        />
+        <h4 className="mb-2">Change Video</h4>
+        <div className="mb-4">
+          <span>
+            YouTube Video ID or URL:
+            <input
+              className="form-control"
+              value={src}
+              onChange={(e) => setSrc(e.target.value)}
+              placeholder="Video ID or URL"
+            />
+          </span>
+        </div>
+
         <h4 className="mb-2">Size</h4>
         <p className="mb-2">
-          You can manually set the image width here and the height will scale automatically.
+          You can manually set the video width here and the height will scale automatically.
         </p>
         <div className="mb-4">
           <span>